requirements.txt
Flask
tensorflow
numpy
pandas
matplotlib
Pillow
sklearn
gunicorn
werkzeug
opencv-python
